# StablingApi/swagger.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "StablingApi",
    "description": "API ИС поддержки управления КСК",
    "version": "v1"
  },
  "paths": {
...
    "/api/BalanceReplenishments": {
      "get": {
        "tags": [
          "BalanceReplenishments"
        ],
        "operationId": "BalanceReplenishments_GetAll",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/BalanceReplenishment"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "BalanceReplenishments"
        ],
        "operationId": "BalanceReplenishments_Create",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "replenishment",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/BalanceReplenishment"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "BalanceReplenishments"
        ],
        "operationId": "BalanceReplenishments_Update",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "replenishment",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/BalanceReplenishment"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/BalanceReplenishments/{id}": {
      "get": {
        "tags": [
          "BalanceReplenishments"
        ],
        "operationId": "BalanceReplenishments_Get",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "guid",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/BalanceReplenishment"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "BalanceReplenishments"
        ],
        "operationId": "BalanceReplenishments_Delete",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "guid",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/api/BalanceReplenishments/ByDate/{date}": {
      "get": {
        "tags": [
          "BalanceReplenishments"
        ],
        "operationId": "BalanceReplenishments_GetByDate",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "date",
            "in": "path",
            "required": true,
            "format": "date-time",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/BalanceReplenishment"
              }
            }
          }
        }
      }
    },
    "/api/BalanceReplenishments/ForShow/ByDate/{date}": {
      "get": {
        "tags": [
          "BalanceReplenishments"
        ],
        "operationId": "BalanceReplenishments_GetForShowByDate",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "date",
            "in": "path",
            "required": true,
            "format": "date-time",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/BalanceReplenishmentForShow"
              }
            }
          }
        }
      }
    },
    "/api/BalanceReplenishments/ForShow/{id}": {
      "get": {
        "tags": [
          "BalanceReplenishments"
        ],
        "operationId": "BalanceReplenishments_GetForShow",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "guid",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/BalanceReplenishmentForShow"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "BalanceReplenishment": {
      "type": "object",
      "required": [
        "balanceReplenishmentId",
        "clientId",
        "trainerId",
        "replenishmentDate",
        "amount"
      ],
      "properties": {
        "balanceReplenishmentId": {
          "type": "string",
          "format": "guid"
        },
        "clientId": {
          "type": "integer",
          "format": "int32"
        },
        "trainerId": {
          "type": "integer",
          "format": "int32"
        },
        "replenishmentDate": {
          "type": "string",
          "format": "date-time"
        },
        "amount": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "BalanceReplenishmentForShow": {
      "type": "object",
      "required": [
        "balanceReplenishmentId",
        "trainerName",
        "clientName",
        "amount",
        "replenishmentDate"
      ],
      "properties": {
        "balanceReplenishmentId": {
          "type": "string",
          "format": "guid"
        },
        "trainerName": {
          "type": "string"
        },
        "clientName": {
          "type": "string"
        },
        "amount": {
          "type": "number",
          "format": "double"
        },
        "replenishmentDate": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
  }
}